Improve documentation
