+++ Sources/Player/Player+Navigation.swift
@@ -4,80 +4,75 @@
 //  License information is available from the LICENSE file.
 //
 
-import AVFoundation
+import CoreMedia
 import DequeModule
+import Foundation
 
 public extension Player {
-    /// Checks whether returning to the previous item in the queue is possible.
+    /// Checks whether returning to the previous content is possible.
     ///
     /// - Returns: `true` if possible.
-    func canReturnToPreviousItem() -> Bool {
-        canReturnToItem(before: currentIndex, in: storedItems)
-    }
-
-    /// Returns to the previous item in the queue.
     ///
-    /// Skips failed items.
-    func returnToPreviousItem() {
-        guard canReturnToPreviousItem() else { return }
-        queuePlayer.replaceItems(with: AVPlayerItem.playerItems(from: returningItems, length: configuration.preloadedItems, reload: true))
+    /// > Important: Observes the ``PlayerConfiguration/navigationMode`` set in the ``Player/configuration``.
+    func canReturnToPrevious() -> Bool {
+        canReturn(before: currentIndex, in: storedItems, streamType: streamType)
     }
 
-    /// Checks whether moving to the next item in the queue is possible.
+    /// Returns to the previous content.
+    ///
+    /// > Important: Observes the ``PlayerConfiguration/navigationMode`` set in the ``Player/configuration``.
+    func returnToPrevious() {
+        if shouldSeekToStartTime() {
+            seek(near(.zero))
+        }
+        else {
+            returnToPreviousItem()
+        }
+    }
+
+    /// Checks whether moving to the next content is possible.
     ///
     /// - Returns: `true` if possible.
-    func canAdvanceToNextItem() -> Bool {
-        canAdvanceToItem(after: currentIndex, in: storedItems)
-    }
-
-    /// Moves to the next item in the queue.
-    func advanceToNextItem() {
-        guard canAdvanceToNextItem() else { return }
-        queuePlayer.replaceItems(with: AVPlayerItem.playerItems(from: advancingItems, length: configuration.preloadedItems, reload: true))
-    }
-
-    /// Makes the item at the specified index become the current one.
     ///
-    /// - Parameter index: The item index.
-    func setCurrentIndex(_ index: Int) throws {
-        guard (0..<storedItems.count).contains(index) else { throw PlaybackError.itemOutOfBounds }
-        let playerItems = AVPlayerItem.playerItems(from: Array(storedItems.suffix(from: index)), length: configuration.preloadedItems, reload: true)
-        queuePlayer.replaceItems(with: playerItems)
+    /// > Important: Observes the ``PlayerConfiguration/navigationMode`` set in the ``Player/configuration``.
+    func canAdvanceToNext() -> Bool {
+        canAdvanceToNextItem()
+    }
+
+    /// Moves to the next content.
+    ///
+    /// > Important: Observes the ``PlayerConfiguration/navigationMode`` set in the ``Player/configuration``.
+    func advanceToNext() {
+        advanceToNextItem()
     }
 }
 
 extension Player {
-    func canReturnToItem(before index: Int?, in items: Deque<PlayerItem>) -> Bool {
-        !Self.items(before: index, in: items).isEmpty
+    func canReturn(before index: Int?, in items: Deque<PlayerItem>, streamType: StreamType) -> Bool {
+        switch configuration.navigationMode {
+        case .smart where streamType == .onDemand:
+            return true
+        default:
+            return canReturnToItem(before: index, in: items)
+        }
     }
 
-    func canAdvanceToItem(after index: Int?, in items: Deque<PlayerItem>) -> Bool {
-        !Self.items(after: index, in: items).isEmpty
+    func canAdvance(after index: Int?, in items: Deque<PlayerItem>) -> Bool {
+        canAdvanceToItem(after: index, in: items)
     }
 }
 
 private extension Player {
-    /// Returns the list of items to be loaded to return to the previous (playable) item.
-    var returningItems: [PlayerItem] {
-        Self.items(before: currentIndex, in: storedItems)
+    func isAwayFromStartTime(withInterval interval: TimeInterval) -> Bool {
+        time.isValid && seekableTimeRange.isValid && (time - seekableTimeRange.start).seconds >= interval
     }
 
-    /// Returns the list of items to be loaded to advance to the next (playable) item.
-    var advancingItems: [PlayerItem] {
-        Self.items(after: currentIndex, in: storedItems)
-    }
-
-    static func items(before index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
-        guard let index else { return [] }
-        let previousIndex = items.index(before: index)
-        guard previousIndex >= 0 else { return [] }
-        return Array(items.suffix(from: previousIndex))
-    }
-
-    static func items(after index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
-        guard let index else { return [] }
-        let nextIndex = items.index(after: index)
-        guard nextIndex < items.count else { return [] }
-        return Array(items.suffix(from: nextIndex))
+    func shouldSeekToStartTime() -> Bool {
+        switch configuration.navigationMode {
+        case .immediate:
+            return false
+        case let .smart(interval: interval):
+            return (streamType == .onDemand && isAwayFromStartTime(withInterval: interval)) || !canReturnToPreviousItem()
+        }
     }
 }
